// src/types/carValueInfo.ts
export interface CarValueInfo {
  id: number;
  model: string;
  year: number;
}

// export let carValues: CarValueInfo[] = [
//   {
//     id: 1,
//     model: "civic",
//     year: 2014,
//   },
//   {
//     id: 2,
//     model: "accord",
//     year: 2015,
//   },
//   {
//     id: 3,
//     model: "sonata",
//     year: 2013,
//   },
//   {
//     id: 4,
//     model: "camry",
//     year: 2011,
//   },
//   {
//     id: 5,
//     model: "model 3",
//     year: 2022,
//   },
//   {
//     id: 6,
//     model: "911",
//     year: 2020,
//   },
//   {
//     id: 7,
//     model: "altima*",
//     year: 2019,
//   },
//   {
//     id: 8,
//     model: "a4",
//     year: 2018,
//   },
// ];